\name{hollisterFig}
\alias{hollisterFig}
\title{Create Hollister Figure}
\usage{
hollisterFig(doi)
}
\arguments{
  \item{doi}{The doi for the Hollister et al. 2011 paper.
  Use this to retrieve article level metrics}
}
\description{
This function generates the figure showing article level
metrics for Hollister, 2011.
}
\references{
Hollister JW, Milstead WB, Urrutia MA (2011) Predicting
Maximum Lake Depth from Surrounding Topography. PLoS ONE
6(9): e25764. doi:10.1371/journal.pone.0025764
}

