\name{milsteadFig}
\alias{milsteadFig}
\title{Create Milstead figure}
\usage{
milsteadFig(doi)
}
\arguments{
  \item{doi}{The doi for the Milstead et al. 2013 paper.
  Use this to retrieve article level metrics}
}
\description{
This function generates the figure showing article level
metrics for Milstead, 2013.
}
\references{
Milstead WB, Hollister JW, Moore RB, Walker HA (2013)
Estimating Summer Nutrient Concentrations in Northeastern
Lakes from SPARROW Load Predictions and Modeled Lake Depth
and Volume. PLoS ONE 8(11): e81457.
doi:10.1371/journal.pone.0081457
}

